feat: move  mergeDefaultAudioOutput to libs/common
