+++ libs/common/srcBrowser/helpers/index.ts
@@ -1,2 +1,4 @@
 export * from '../../src/helpers';
 export { default as composeProviders, type ProviderComponent } from './composeProviders';
+export { default as mergeDefaultAudioOutput } from './mergeDefaultAudioOutput';
+export type { MergeDefaultAudioOutputResult } from './mergeDefaultAudioOutput';
